requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
Pillow==10.4.0
twilio==9.3.4
africastalking==1.2.6
africas-talking==1.2.3
reportlab==4.2.5
openpyxl==3.1.5
gunicorn==23.0.0